feat: 上传文件接口写完

+++ packages/gin_series/06_upload_file/model/usreModel.go
@@ -7,7 +7,7 @@ import (
 )
 
 type UserModel struct {
-	Id       int64          `form:"id"`
+	Id       int            `form:"id"`
 	Email    string         `form:"email" binding:"email"`
 	Password string         `form:"password"`
 	Avatar   sql.NullString `form:"avatar"`
